+++ Examples/distributed_training.py
@@ -1,89 +1,39 @@
-from xcube.core.store import new_data_store
-from global_land_mask import globe
-import dask.array as da
-import xarray as xr
-import random
-from torch import nn
 import torch
 import numpy as np
+import xarray as xr
+import dask.array as da
+from torch import nn
+from global_land_mask import globe
 from torch.utils.data import TensorDataset
+from xcube.core.store import new_data_store
+from ml4xcube.datasets.xr_dataset import XrDataset
+from ml4xcube.cube_utilities import get_chunk_sizes
+from ml4xcube.data_assignment import assign_block_split
+from ml4xcube.statistics import standardize, get_statistics
+from ml4xcube.datasets.pytorch_xr import prepare_dataloader
+from ml4xcube.training.pytorch_distributed import ddp_init, dist_train, Trainer
 
-# add path, if mltools not installed
-import sys
-sys.path.append('../mltools')
 
-from mltools.cube_utilities import get_chunk_by_index, get_chunk_sizes, calculate_total_chunks
-from mltools.distributed_training import ddp_init, prepare_dataloader, dist_train, Trainer
-from mltools.data_assignment import assign_block_split
-from mltools.statistics import standardize
-
-from mltools.geo_plots import plot_geo_data
+# To utilize ml4xcube for distributed training, use the following command with torchrun to initiate the process:
+#
+# torchrun --standalone --nproc_per_node=<number_of_processes> distributed_training.py <epochs>
+#
+# Replace `<number_of_processes>` with the number of processes you wish to run per node,
+# and `<epochs>` with the total number of training epochs.
 
 
 def preprocess_data(ds: xr.Dataset):
-    """Preprocess and split dataset into training and testing datasets.
+    ds = XrDataset(ds, 3).get_dataset()
 
-    Args:
-    ds (xr.Dataset): The input dataset to be processed.
+    at_stat = get_statistics(ds, 'air_temperature_2m')
+    lst_stat = get_statistics(ds, 'land_surface_temperature')
 
-    Returns:
-    Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]: Returns preprocessed and split training and testing data.
-    """
+    X = standardize(ds['air_temperature_2m'], *at_stat)
+    y = standardize(ds['land_surface_temperature'], *lst_stat)
 
-    # Calculate the total number of data chunks in the dataset
-    total_chunks = calculate_total_chunks(ds)
-
-    # Initialize lists to hold the data from selected chunks
-    X_train_all, X_test_all, y_train_all, y_test_all = [], [], [], []
-
-    # Keep track of processed chunks to avoid repetition
-    processed_chunks = list()
-
-    # Process chunks until 3 unique chunks have been processed
-    while len(processed_chunks) < 3:
-        chunk_index = random.randint(0, total_chunks - 1)  # Select a random chunk index
-        if chunk_index in processed_chunks:
-            continue  # Skip if this chunk has already been processed
-
-        # Retrieve the chunk by its index
-        chunk = get_chunk_by_index(ds, chunk_index)
-
-        # Flatten the data and select only land values, then drop NaN values
-        cf = {x: chunk[x].ravel() for x in chunk.keys()}
-        lm = cf['land_mask']
-        cft = {x: cf[x][lm == True] for x in cf.keys()}
-        lst = cft['land_surface_temperature']
-        cfn = {x: cft[x][~np.isnan(lst)] for x in cf.keys()}
-
-        # Proceed only if there are valid land surface temperature data points
-        if len(cfn['land_surface_temperature']) > 0:
-            processed_chunks.append(chunk_index)
-            X = cfn['air_temperature_2m']
-            y = cfn['land_surface_temperature']
-
-            # Split the data based on the 'split' attribute
-            X_train, X_test = X[cfn['split'] == True], X[cfn['split'] == False]
-            y_train, y_test = y[cfn['split'] == True], y[cfn['split'] == False]
-
-            # Append processed data to their respective lists
-            X_train_all.append(X_train)
-            X_test_all.append(X_test)
-            y_train_all.append(y_train)
-            y_test_all.append(y_test)
-
-    # Concatenate the data from all processed chunks
-    X_train = np.concatenate(X_train_all)
-    X_test = np.concatenate(X_test_all)
-    y_train = np.concatenate(y_train_all)
-    y_test = np.concatenate(y_test_all)
-
-    # Standardize the data
-    X_mean, X_std = np.mean(X_train), np.std(X_train)
-    y_mean, y_std = np.mean(y_train), np.std(y_train)
-    X_train = standardize(X_train, X_mean, X_std)
-    X_test = standardize(X_test, X_mean, X_std)
-    y_train = standardize(y_train, y_mean, y_std)
-    y_test = standardize(y_test, y_mean, y_std)
+    # Split the data based on the 'split' attribute
+    X_train, X_test = X[ds['split'] == True], X[ds['split'] == False]
+    y_train, y_test = y[ds['split'] == True], y[ds['split'] == False]
 
     return X_train, X_test, y_train, y_test
 
@@ -127,8 +77,8 @@ def main(save_every: int, total_epochs: int, batch_size: int, snapshot_path: str
 
     # Load training objects and prepare data loaders
     train_set, test_set, model, optimizer = load_train_objs()
-    train_loader = prepare_dataloader(train_set, batch_size, num_workers=5)
-    test_loader = prepare_dataloader(test_set, batch_size, num_workers=5)
+    train_loader = prepare_dataloader(train_set, batch_size, num_workers=5, parallel=True)
+    test_loader = prepare_dataloader(test_set, batch_size, num_workers=5, parallel=True)
 
     # Initialize the trainer and start training
     trainer = Trainer(model, train_loader, test_loader, optimizer, save_every, best_model_path, snapshot_path, task_type='supervised')